перенес логику из активити во фрагмент
