+++ app/src/main/java/com/example/notesapp/ui/NotesListActivity.java
@@ -18,11 +18,7 @@ import com.example.notesapp.data.Repo;
 import com.example.notesapp.data.InMemoryRepoImpl;
 import com.example.notesapp.recycler.NotesAdapter;
 
-public class NotesListActivity extends AppCompatActivity implements NotesAdapter.onNoteClickListener {
-
-    private Repo repository = InMemoryRepoImpl.getInstance();
-    private RecyclerView recyclerView;
-    private NotesAdapter adapter;
+public class NotesListActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -34,61 +30,5 @@ public class NotesListActivity extends AppCompatActivity implements NotesAdapter
                 .beginTransaction()
                 .replace(R.id.fragment_container, notesListFragment)
                 .commit();
-
-        if (repository.getAll().size() == 0) fillRepo();
-
-        adapter = new NotesAdapter();
-        adapter.setNotes(repository.getAll());
-
-        adapter.setOnNoteClickListener(this);
-
-        recyclerView = findViewById(R.id.rv_list);
-        recyclerView.setLayoutManager(new LinearLayoutManager(this));
-        recyclerView.setAdapter(adapter);
-    }
-
-    private void fillRepo() {
-        repository.create(new Note("Title 1", "Description 1"));
-        repository.create(new Note("Title 2", "Description 2"));
-        repository.create(new Note("Title 3", "Description 3"));
-        repository.create(new Note("Title 4", "Description 4"));
-        repository.create(new Note("Title 5", "Description 5"));
-        repository.create(new Note("Title 6", "Description 6"));
-        repository.create(new Note("Title 7", "Description 7"));
-        repository.create(new Note("Title 8", "Description 8"));
-        repository.create(new Note("Title 9", "Description 9"));
-        repository.create(new Note("Title 10", "Description 10"));
-        repository.create(new Note("Title 11", "Description 11"));
-        repository.create(new Note("Title 12", "Description 12"));
-        repository.create(new Note("Title 13", "Description 13"));
-        repository.create(new Note("Title 14", "Description 14"));
-        repository.create(new Note("Title 15", "Description 15"));
-        repository.create(new Note("Title 16", "Description 16"));
-    }
-
-    @Override
-    public void onNoteClick(Note note) {
-        Intent intent = new Intent(this, EditNoteActivity.class);
-        intent.putExtra(Constants.NOTE, note);
-        startActivity(intent);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.main, menu);
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    @SuppressLint("NonConstantResourceId")
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.main_create:
-                Intent intent = new Intent(this, EditNoteActivity.class);
-                startActivity(intent);
-                return true;
-        }
-        return super.onOptionsItemSelected(item);
-
     }
 }
